+++ examples/find_faces_in_picture.py
@@ -4,7 +4,9 @@ import face_recognition
 # Load the jpg file into a numpy array
 image = face_recognition.load_image_file("biden.jpg")
 
-# Find all the faces in the image
+# Find all the faces in the image using the default HOG-based model.
+# This method is fairly accurate, but not as accurate as the CNN model and not GPU accelerated.
+# See also: find_faces_in_picture_cnn.py
 face_locations = face_recognition.face_locations(image)
 
 print("I found {} face(s) in this photograph.".format(len(face_locations)))