+++ face_recognition/api.py
@@ -17,6 +17,9 @@ face_detector = dlib.get_frontal_face_detector()
 predictor_model = face_recognition_models.pose_predictor_model_location()
 pose_predictor = dlib.shape_predictor(predictor_model)
 
+cnn_face_detection_model = face_recognition_models.cnn_face_detector_model_location()
+cnn_face_detector = dlib.cnn_face_detection_model_v1(cnn_face_detection_model)
+
 face_recognition_model = face_recognition_models.face_recognition_model_location()
 face_encoder = dlib.face_recognition_model_v1(face_recognition_model)
 
@@ -78,26 +81,66 @@ def load_image_file(file, mode='RGB'):
     return scipy.misc.imread(file, mode=mode)
 
 
-def _raw_face_locations(img, number_of_times_to_upsample=1):
+def _raw_face_locations(img, number_of_times_to_upsample=1, model="hog"):
     """
     Returns an array of bounding boxes of human faces in a image
 
     :param img: An image (as a numpy array)
+    :param number_of_times_to_upsample: How many times to upsample the image looking for faces. Higher numbers find smaller faces.
+    :param model: Which face detection model to use. "hog" is less accurate but faster on CPUs. "cnn" is a more accurate
+                  deep-learning model which is GPU/CUDA accelerated (if available). The default is "hog".
+    :return: A list of dlib 'rect' objects of found face locations
+    """
+    if model == "cnn":
+        return cnn_face_detector(img, number_of_times_to_upsample)
+    else:
+        return face_detector(img, number_of_times_to_upsample)
+
+
+def face_locations(img, number_of_times_to_upsample=1, model="hog"):
+    """
+    Returns an array of bounding boxes of human faces in a image
+
+    :param img: An image (as a numpy array)
+    :param number_of_times_to_upsample: How many times to upsample the image looking for faces. Higher numbers find smaller faces.
+    :param model: Which face detection model to use. "hog" is less accurate but faster on CPUs. "cnn" is a more accurate
+                  deep-learning model which is GPU/CUDA accelerated (if available). The default is "hog".
+    :return: A list of tuples of found face locations in css (top, right, bottom, left) order
+    """
+    if model == "cnn":
+        return [_trim_css_to_bounds(_rect_to_css(face.rect), img.shape) for face in _raw_face_locations(img, number_of_times_to_upsample, "cnn")]
+    else:
+        return [_trim_css_to_bounds(_rect_to_css(face), img.shape) for face in _raw_face_locations(img, number_of_times_to_upsample, model)]
+
+
+def _raw_face_locations_batched(images, number_of_times_to_upsample=1):
+    """
+    Returns an 2d array of dlib rects of human faces in a image using the cnn face detector
+
+    :param img: A list of images (each as a numpy array)
     :param number_of_times_to_upsample: How many times to upsample the image looking for faces. Higher numbers find smaller faces.
     :return: A list of dlib 'rect' objects of found face locations
     """
-    return face_detector(img, number_of_times_to_upsample)
+    return cnn_face_detector(images, number_of_times_to_upsample)
 
 
-def face_locations(img, number_of_times_to_upsample=1):
+def batch_face_locations(images, number_of_times_to_upsample=1, batch_size=128):
     """
-    Returns an array of bounding boxes of human faces in a image
+    Returns an 2d array of bounding boxes of human faces in a image using the cnn face detector
+    If you are using a GPU, this can give you much faster results since the GPU
+    can process batches of images at once. If you aren't using a GPU, you don't need this function.
 
-    :param img: An image (as a numpy array)
+    :param img: A list of images (each as a numpy array)
     :param number_of_times_to_upsample: How many times to upsample the image looking for faces. Higher numbers find smaller faces.
+    :param batch_size: How many images to include in each GPU processing batch.
     :return: A list of tuples of found face locations in css (top, right, bottom, left) order
     """
-    return [_trim_css_to_bounds(_rect_to_css(face), img.shape) for face in _raw_face_locations(img, number_of_times_to_upsample)]
+    def convert_cnn_detections_to_css(detections):
+        return [_trim_css_to_bounds(_rect_to_css(face.rect), images[0].shape) for face in detections]
+
+    raw_detections_batched = _raw_face_locations_batched(images, number_of_times_to_upsample)
+
+    return list(map(convert_cnn_detections_to_css, raw_detections_batched))
 
 
 def _raw_face_landmarks(face_image, face_locations=None):